added initial new normalise, only replicates filter
